+++ admin/elevation/sgpx.php
@@ -63,9 +63,11 @@ function leafext_form_sgpx($field) {
 
 // Sanitize and validate input. Accepts an array, return a sanitized array.
 function leafext_validate_sgpx_options($options) {
-	if (isset($_POST['submit'])) return $options;
-	if (isset($_POST['delete'])) delete_option('leafext_sgpxparams');
-	return false;
+	if ( ! empty( $_POST ) && check_admin_referer( 'leafext_elevation', 'leafext_elevation_nonce' ) ) {
+		if (isset($_POST['submit'])) return $options;
+		if (isset($_POST['delete'])) delete_option('leafext_sgpxparams');
+		return false;
+	}
 }
 
 // init sgpx_unclean_db
@@ -83,16 +85,18 @@ function leafext_form_sgpx_unclean_db() {
 
 // Submit sgpx_unclean_db
 function leafext_validate_sgpx_unclean_db($input) {
+	if ( ! empty( $_POST ) && check_admin_referer( 'leafext_elevation', 'leafext_elevation_nonce' ) ) {
 		if (isset($_POST['delete'])) {
-		if ( $_POST['delete'] == __( 'Delete all settings from wp-gpx-maps!', 'extensions-leaflet-map' ) ) {
-			global $wpdb;
-			$option_names = $wpdb->get_results( "SELECT option_name FROM $wpdb->options WHERE option_name LIKE 'wpgpxmaps_%' " );
-			foreach ( $option_names as $key=>$value ) {
-				delete_option($value->option_name);
+			if ( $_POST['delete'] == __( 'Delete all settings from wp-gpx-maps!', 'extensions-leaflet-map' ) ) {
+				global $wpdb;
+				$option_names = $wpdb->get_results( "SELECT option_name FROM $wpdb->options WHERE option_name LIKE 'wpgpxmaps_%' " );
+				foreach ( $option_names as $key=>$value ) {
+					delete_option($value->option_name);
+				}
 			}
 		}
+		return false;
 	}
-	return false;
 }
 
 // Helptext
